+++ core-libs/tach-acid-impulse-lib/src/Tach/Acid/Impulse/State.hs
@@ -3,7 +3,7 @@
   , TemplateHaskell, TypeFamilies, RecordWildCards, DeriveGeneric, DeriveDataTypeable #-}
 
 
-module Tach.Migration.Acidic.State where
+module Tach.Acid.Impulse.State where
 
 -- import Tach.Migration.Acidic.State.Internal
 import Tach.Migration.Acidic.Types
@@ -19,10 +19,11 @@ import qualified Data.Vector as V
 
 -- Intial State defnitions
 
--- | This is a simple impulse valued sequence.  No Compression, No structural change.
--- Note, the terrible type signature occurs because safecopy hates type synonyms!
--- newtype TVSimpleImpulseTypeStore = TVSimpleImpulseTypeStore { unTimeValueStore :: (ImpulseSeries (ImpulsePeriod (Vector Double) Integer ) (ImpulseStart Integer) (ImpulseEnd Integer) (ImpulseRep (Seq TVSimple))) } deriving (Typeable,Generic)
 
+
+
+initialImpulseKey :: TVKey
+initialImpulseKey = ImpulseKey 0
 
 initialImpulseStart :: TVSStart -- ImpulseStart Integer
 initialImpulseStart       = ImpulseStart 0
@@ -30,7 +31,7 @@ initialImpulseStart       = ImpulseStart 0
 initialImpulseEnd :: TVSEnd -- ImpulseEnd Integer
 initialImpulseEnd       = ImpulseEnd 0
 
-initialImpulseRep :: ImpulseRep (S.Seq TVSimple)
+initialImpulseRep :: ImpulseRep (S.Seq TVNoKey)
 initialImpulseRep       = ImpulseRep (S.empty)
 
 -- | IntiialTVSimple not needed cause the rep is empty
@@ -40,8 +41,10 @@ initialImpulseRep       = ImpulseRep (S.empty)
 
 initialImpulsePeriod :: TVPeriod -- ImpulsePeriod (V.Vector Double) (Integer)
 initialImpulsePeriod       = IPeriodParameterized (V.empty)
+
+
 initialImpulseSeries ::     TVSimpleImpulseType
-initialImpulseSeries       = ImpulseSeries initialImpulsePeriod initialImpulseStart initialImpulseEnd initialImpulseRep 
+initialImpulseSeries       = ImpulseSeries initialImpulseKey initialImpulsePeriod initialImpulseStart initialImpulseEnd initialImpulseRep 
 
 initialTVSimpleImpulseTypeStore :: TVSimpleImpulseTypeStore 
 initialTVSimpleImpulseTypeStore       = TVSimpleImpulseTypeStore initialImpulseSeries