+++ core-libs/tach-acid-impulse-lib/src/Tach/Acid/Impulse/State/Internal.hs
@@ -1,0 +1,3 @@
+module Tach.Acid.Impulse.State.Internal
+    (
+    ) where