+++ core-libs/tach-migration-acid-lib/src/Tach/Migration/Acidic/State/Internal.hs
@@ -1,3 +1,0 @@
-module Tach.Migration.Acidic.State.Internal
-    (
-    ) where