+++ core-libs/tach-acid-impulse-lib/changelog.md
@@ -11,11 +11,11 @@ Change Log
 
 *For complete history please view commit log*
 
-##Current Version 
+##Current Version 0.1.1
 --------------------------	
 
 
-### Proposed Release 0.1.1
+### Proposed Release 0.1.2
 ----------------------------
 
 
@@ -23,6 +23,6 @@ Change Log
 *Change History* 
 
 + 0.1.1
-
+  Changed the directory structure around ... New : Tach/Acid/Impulse
 + 0.1.0
   Initial Commit