+++ core-types/tach-migration-acidic/changelog.md
@@ -11,17 +11,18 @@ Change Log
 
 *For complete history please view commit log*
 
-##Current Version  0.1.1
+##Current Version  0.1.2
 --------------------------	
 
 
-### Proposed Release 0.1.2
+### Proposed Release 0.1.3
 ----------------------------
 
 
 
 *Change History* 
-
++ 0.1.2
+  updated impulse types version
 + 0.1.1
   added Key to impulse series
 + 0.1.0